requirements.txt
Flask==2.2.2
line-bot-sdk
openai
APScheduler
Flask-SQLAlchemy